requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
typing_extensions==4.12.2
urllib3==2.2.3
